refactor(conversation):WB-3957 rename MessageList.tsx
