+++ conversation/frontend/src/features/message-list/MessageList.test.tsx
@@ -1,7 +1,7 @@
 import { act } from 'react';
 import { render, renderHook, screen, waitFor, wrapper } from '~/mocks/setup';
 import { useFolderMessages } from '~/services';
-import { MessageList } from '.';
+import { MessageList } from './MessageList';
 
 /**
  * Mock useParams
